+++ spec/chef/recipes/gitlab-rails_spec.rb
@@ -456,6 +456,23 @@ describe 'gitlab::gitlab-rails' do
         end
       end
 
+      context 'when geo prune event log worker is configured' do
+        it 'sets the cron value' do
+          stub_gitlab_rb(gitlab_rails: { geo_prune_event_log_worker_cron: '5 4 3 2 1' })
+
+          expect(chef_run).to render_file(gitlab_yml_path)
+            .with_content(/geo_prune_event_log_worker:\s+cron:\s+"5 4 3 2 1"/)
+        end
+      end
+
+      context 'when geo prune event log worker is not configured' do
+        it 'does not set the cron value' do
+          expect(chef_run).to render_file(gitlab_yml_path).with_content { |content|
+            expect(content).not_to include('geo_prune_event_log_worker')
+          }
+        end
+      end
+
       context 'when file download dispatch worker is configured' do
         it 'sets the cron value' do
           stub_gitlab_rb(gitlab_rails: { geo_file_download_dispatch_worker_cron: '1 2 3 4 5' })