+++ CHANGELOG.md
@@ -8,6 +8,7 @@ omnibus-gitlab repository.
 - Disable PgBouncer PID file by default
 - Add pgbouncer_exporter
 - Bump minimum version required for upgrade to 10.8 2df263267
+- Geo: Only create database_geo.yml and run migrations if Rails is enabled
 - Render gitlab-pages admin settings
 - Fix old unicorn master not quitting after a new process is running
 - Mattermost: Fix reconfiguration of GitLab OAuth configuration settings
@@ -21,6 +22,8 @@ omnibus-gitlab repository.
 - Bump git to 2.17.1 !2552
 - Updated Mattermost to 4.10.1
 - Render gitaly-ruby num_workers config
+- Set installation type information to be used by usage ping - !2561
+- Add graphicsmagick dependency
 
 10.8.2